[src] Various mostly-cosmetic changes (copying from another branch) (#3109)
